requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Pillow==10.2.0
pdf2image==1.16.0
img2pdf==0.4.4
python-dotenv==0.19.0
Flask-WTF==0.15.1
stripe==7.14.0
werkzeug==3.0.1
PyJWT>=2.5.0
python-dateutil==2.8.2
firebase-admin==6.4.0
pyrebase4==4.7.1
